Edit New issue Opt-in NuGet support for new project types #569
CPS projects are now checked for project capabilities. NuGet is made
available when a CPS project has these capabilities defined:

Project Capability | Description
------------ | -------------
CPS | This is a "pure" CPS based project. Visual C++ does not declare
this capability, but all other CPS projects do.
AssemblyReferences | Indicates that the project supports assembly
references (distinct from WinRTReferences)
DeclaredSourceItems | Indicates that the project is a typical MSBuild
project (not DNX) in that it declares source items in the project itself
(rather than a project.json file that assumes all files in the directory
are part of a compilation).
UserSourceItems | Indicates that the user is allowed to add arbitrary
files to their project.