Remove CPS as a required project capability
NuGet requires "CPS" to be present in the project capability expression before it considers it OK to interact with. But any compliant project system can signify that they support references and source items via project capabilities and thereby work with NuGet without adding to the project type guid "whitelist" that the NuGet VS extension has.

This change removes "CPS" as a required project capability to facilitate this.